simplify region selection algo

+++ src/main/scala/cn/graiph/blobfs/messages.scala
@@ -3,13 +3,13 @@ package cn.graiph.blobfs
 /**
   * Created by bluejoe on 2019/8/23.
   */
-case class CreateRegionRequest(regionId: Int) {
+case class CreateRegionRequest(regionId: Long) {
 }
 
-case class CreateRegionResponse(regionId: Int) {
+case class CreateRegionResponse(regionId: Long) {
 }
 
-case class SendCompleteFileRequest(regionId: Option[Int], bytes: Array[Byte], totalLength: Long) {
+case class SendCompleteFileRequest(regionId: Option[Long], bytes: Array[Byte], totalLength: Long) {
 
 }
 
@@ -20,7 +20,7 @@ case class DiscardChunksRequest(transId: Long) {
 
 }
 
-case class StartSendChunksRequest(regionId: Option[Int], totalLength: Long) {
+case class StartSendChunksRequest(regionId: Option[Long], totalLength: Long) {
 
 }
 
@@ -32,6 +32,6 @@ case class SendChunkRequest(transId: Long, chunkBytes: Array[Byte], offset: Long
 
 }
 
-case class SendChunkResponse(fileId: Option[FileId], chunkLength: Int) {
+case class SendChunkResponse(fileId: Option[FileId], chunkLength: Long) {
 
 }
